pyproject.toml
[project]
name = "karton-capa"
version = "0.1.0"
description = "Karton service to run CAPA on samples"
readme = "README.md"
authors = [{ name = "Kevin Ziadeh" }]
requires-python = ">=3.12"
dependencies = [
    "flare-capa>=9.2.1",
    "karton-core>=5.6.0",
]
classifiers = [ 
    "Private :: Do not Upload",
    "Operating System :: Unix",
    "Intended Audience :: Developers",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.12",
]

[tool.ruff]
line-length = 120

[tool.ruff.lint]
select = ["ALL"]
ignore = [
    "S603", # subprocess.check_output
    "G004", # f-string log strings
    "TRY003", # long strings outside exception class
    "EM101", # strings in error raise
    "D100", # missing docstring in public module
    "D104", # missing docstring in public package
    "D212", # multi-line docstring summary should start at the first line
]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["E402", "N999"]

[project.scripts]
karton-capa = "src.capa:Capa.main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]

[dependency-groups]
dev = [
    "ruff>=0.9.8",
]
